tensorflow_ml/requirements.txt
numpy
tensorflow
tensorflow_recommenders
tensorflow_datasets
tensorflow_io
kafka-python
pandas
flask 
flask-cors

